Move getTooltipContent to its own file
